requirements.txt
aioredis==1.3.1
aiosqlite==0.20.0
alembic==1.13.3
amqp==5.3.1
annotated-types==0.7.0
anyio==4.9.0
async-timeout==5.0.1
asyncpg==0.30.0
bcrypt==4.0.1
billiard==4.2.1
black==25.1.0
celery==5.5.0
certifi==2025.1.31
click==8.1.8
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
dnspython==2.7.0
ecdsa==0.19.1
email_validator==2.2.0
fastapi==0.115.0
fastapi-cache==0.1.0
fastapi-cli==0.0.7
fastapi-redis-cache==0.2.5
greenlet==3.1.1
gunicorn==23.0.0
h11==0.14.0
hiredis==3.1.0
httpcore==1.0.7
httptools==0.6.4
httpx==0.28.1
idna==3.10
isort==6.0.1
itsdangerous==2.2.0
Jinja2==3.1.4
kombu==5.5.2
loguru==0.7.2
Mako==1.3.9
markdown-it-py==3.0.0
MarkupSafe==3.0.2
mdurl==0.1.2
mypy-extensions==1.0.0
orjson==3.10.16
packaging==24.2
passlib==1.7.4
pathspec==0.12.1
platformdirs==4.3.7
prompt_toolkit==3.0.50
pyasn1==0.6.1
pydantic==2.9.2
pydantic-extra-types==2.10.3
pydantic-settings==2.5.2
pydantic_core==2.23.4
Pygments==2.19.1
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
python-jose==3.3.0
python-multipart==0.0.20
PyYAML==6.0.2
redis==5.2.1
rich==14.0.0
rich-toolkit==0.14.1
rsa==4.9
shellingham==1.5.4
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.35
starlette==0.38.6
typer==0.15.2
typing_extensions==4.13.0
tzdata==2025.2
ujson==5.10.0
uvicorn==0.31.0
uvloop==0.21.0
vine==5.1.0
watchfiles==1.0.4
wcwidth==0.2.13
websockets==15.0.1
